added input data methods and encoding

+++ backend/main.py
@@ -25,12 +25,36 @@ treenodes = db.treenodes
 attributes = db.attribute
 applications = db.antraege
 
-def generate_tree():
-    # load applications list from database
-    application_list = list(applications.find())
+# create path to local assets
+file_dir = Path(__file__)
+dir = file_dir.parent
 
-    # load attribute list from database
-    attribute_list = list(attributes.find({},{"Name":1,"_id":0,"Kategorie":1}))
+def generate_tree():
+
+    # asks for the method of input
+    insertionmode = input("Should the input data for the algorithm come from the database? [Y/n]")
+
+    if insertionmode == "Y":
+        # load applications list from database
+        application_list = list(applications.find())
+        # load applications list from database
+        attribute_list = list(attributes.find({},{"Name":1,"_id":0,"Kategorie":1}))
+
+    elif insertionmode == "n":
+        # load application list from assets
+        data = open(dir / "Algorithmus/assets/Antraege_countryGroup.json", encoding = 'utf-8')
+        application_list = json.load(data)
+        # load attribute list from assets
+        data = open(dir / "Algorithmus/assets/Attribute_countryGroup.json", encoding = 'utf-8')
+        attribute_list = json.load(data)
+
+    else:
+        print("This was not a valid input.")
+        return
+
+    
+    print("Algorithm in progress...")
+    print("This could take a minute...")
 
     attribute = {}
     for attribut in attribute_list:
@@ -49,15 +73,17 @@ def generate_tree():
     # creates the tree and fills nodelist
     Algorithmus.create_tree(application_list,attribute,[],[], brute_force_depth,nodelist,id,"")
 
-    #! The following two operations directly write (and delete) data in the database. 
-    #! Comment these out if you're only testing the algorithm
-    # delete old nodes
-    treenodes.drop()
-    # insert new nodes
-    treenodes.insert_many(nodelist)
-    
-    # prints the number of nodes
-    print(len(nodelist))
+    print("Algorithm ran successfully and generated " + str(len(nodelist))+ " nodes.")
+
+    # asks if the output should be saved in the database
+    savingmode = input("Do you want to save the nodes in the database? [Y/n]")
+    if savingmode == "Y":
+        # delete old nodes
+        treenodes.drop()
+        # insert new nodes
+        treenodes.insert_many(nodelist)
+
+    print("Done!")
 
 if __name__ == "__main__":
     generate_tree()
